requirements.txt
telebot
pytube
python-dotenv
paramiko
scp